blog_api/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-HTTPAuth
psycopg2-binary
python-dotenv